+++ contentful_management/resource_builder.py
@@ -9,6 +9,7 @@ from .upload import Upload
 from .api_key import ApiKey
 from .webhook import Webhook
 from .snapshot import Snapshot
+from .api_usage import ApiUsage
 from .environment import Environment
 from .ui_extension import UIExtension
 from .content_type import ContentType
@@ -64,6 +65,7 @@ class ResourceBuilder(object):
             'SpaceMembership': SpaceMembership,
             'Organization': Organization,
             'UsagePeriod': UsagePeriod,
+            'ApiUsage': ApiUsage,
             'ApiKey': ApiKey,
             'PreviewApiKey': PreviewApiKey,
             'PersonalAccessToken': PersonalAccessToken,